api/requirements.txt
# FastAPI and ASGI server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-decouple==3.8

# Image processing and AI
opencv-python==4.8.1.78
pillow==10.1.0
pytesseract==0.3.10
numpy==1.24.3
torch==2.1.1
torchvision==0.16.1
transformers==4.35.2

# HTTP client
httpx==0.24.1
aiofiles==23.2.1

# Firebase/Supabase
firebase-admin==6.2.0
supabase==2.0.2

# Utilities
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
isort==5.12.0
flake8==6.1.0

# CORS
fastapi-cors==0.0.6

# File upload
python-multipart==0.0.6

# Logging
structlog==23.2.0

# Rate limiting
slowapi==0.1.9 